[#176924316] Fix renaming + tests
